:truck: Move EventList to own component directory
